Reset isNew flags
Problem was: I was trying to render it second time even before it
finished rendering the first time. Now instead of rendering everything
again I just manipulate DOM properties to remove badge once the tab is
opened and also clearing the blue background. But would have loved if
the background color could be animated away. So instead of updating the
UI right away we can just keep it as it is and when pop up is opened
again it will reflect the new data.